feat: CalendarView 이벤트 props 분리 및 progress 정렬 기능 추가 (#11)
