+++ types/calendar.d.ts
@@ -1,0 +1,12 @@
+import type { Option } from '@/components/Select'
+
+export interface CalendarViewProps {
+  option: Option[]
+  events: CalendarEvent[]
+}
+
+export interface CalendarEvent {
+  title: string
+  date: Date
+  progress?: number
+}
